Refactor location of GIS utilities
