+++ govapp/gis/readers/compression.py
