+++ govapp/gis/readers/utils.py
